+++ CHANGELOG.md
@@ -3561,3 +3561,4 @@
 - [2026-08-11 23:04] fix(navigation): resolve race condition during navigation initialization
 - [2026-08-13 09:01] test(test): implement widget integration tests for test
 - [2026-08-13 09:28] feat(cache): add async cache stream provider
+- [2026-08-13 09:59] feat(ui): integrate custom telemetry logger in ui